+++ src/actions/user.actions.js
@@ -21,24 +21,54 @@ export function clearError () {
   return {type: types.CLEAR_ERROR}
 }
 
+export function loadUser (data) {
+  return dispatch => {
+    let user = {
+      name: data.displayName,
+      email: data.email,
+      photoUrl: data.photoURL,
+      emailVerified: data.emailVerified,
+      uid: data.uid,
+      refreshToken: data.refreshToken,
+      provider: data.providerData.map(p => ({...p}))
+    }
+
+    dispatch(loginSuccess(user))
+    return user
+  }
+}
+
 export function login (email, password) {
   return (dispatch) => {
     return Auth
       .signInWithEmailAndPassword(email, password)
-      .then((data) => {
+      .then(data => loadUser(data)(dispatch))
+  }
+}
+
+export function loginProvider (provider) {
+  return (dispatch) => Auth.signInWithProvider(provider)
+      .then(data => {
+        const profile = data.user
+        const credential = data.credential
+
         let user = {
-          name: data.displayName,
-          email: data.email,
-          photoUrl: data.photoURL,
-          emailVerified: data.emailVerified,
-          uid: data.uid,
-          provider: data.providerData.map(p => ({...p}))
+          name: profile.displayName,
+          email: profile.email,
+          photoUrl: profile.photoURL,
+          emailVerified: profile.emailVerified,
+          uid: profile.uid,
+          refreshToken: profile.refreshToken,
+          provider: profile.providerData.map(p => ({...p})),
+          credential
         }
 
         dispatch(loginSuccess(user))
         return data
       })
-  }
+      .catch(err => {
+        console.info('WE HAVE ERROR!', err)
+      })
 }
 
 export function logout () {