+++ src/utils/ExchangeProvider/CoinmarketcapGetTask.js
@@ -2,13 +2,13 @@ import { GetTask } from '../Networking/GetTask'
 
 export class CoinmarketcapGetTask extends GetTask {
   static fetchRawTask (token) {
-    let url = 'http://coinmarketcap.northpole.ro/api/v5/history/' + token.symbol.toUpperCase() + '_2017.json'
+    let url = 'https://coinmarketcap.northpole.ro/api/v5/history/' + token.symbol.toUpperCase() + '_2017.json'
 
     return new CoinmarketcapGetTask(url)
   }
 
   static fetchCurrentBalanceTask (token) {
-    let url = 'http://coinmarketcap.northpole.ro/api/v5/' + token.symbol.toUpperCase() + '.json'
+    let url = 'https://coinmarketcap.northpole.ro/api/v5/' + token.symbol.toUpperCase() + '.json'
 
 
     return new CoinmarketcapGetTask(url)